remove @Repository from interfaces add @Repository and @Scope to repository implementations implement GameRepository and MatchRepository

+++ rest/src/main/java/com/robocubs4205/cubscout/model/DistrictRepositoryImpl.java
@@ -1,15 +1,20 @@
 package com.robocubs4205.cubscout.model;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Repository;
 
+import javax.jdo.JDOHelper;
 import javax.jdo.PersistenceManager;
 import javax.jdo.PersistenceManagerFactory;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import static org.springframework.beans.factory.config.ConfigurableBeanFactory.SCOPE_PROTOTYPE;
+
 @Repository
+@Scope(SCOPE_PROTOTYPE)
 public class DistrictRepositoryImpl implements DistrictRepository {
 
     private final PersistenceManagerFactory pmf;
@@ -36,7 +41,8 @@ public class DistrictRepositoryImpl implements DistrictRepository {
     @Override
     public void delete(District district) {
         try(PersistenceManager pm = pmf.getPersistenceManager()){
-            pm.deletePersistent(district);
+            if(JDOHelper.isPersistent(district))pm.deletePersistent(district);
+            else pm.newQuery(District.class).filter("this.id=:id").deletePersistentAll(district.getId());
         }
     }
